feat(main): added implemention of servlets for changer

+++ src/main/java/models/ExchangeRate.java
@@ -1,16 +1,18 @@
 package models;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
   private int id;
   private int baseCurrencyId;
   private int targetCurrencyId;
-  private double rate;
+  private BigDecimal rate;
 
   public ExchangeRate() {
 
   }
 
-  public ExchangeRate(int id, int baseCurrencyId, int targetCurrencyId, double rate) {
+  public ExchangeRate(int id, int baseCurrencyId, int targetCurrencyId, BigDecimal rate) {
     this.id = id;
     this.baseCurrencyId = baseCurrencyId;
     this.targetCurrencyId = targetCurrencyId;
@@ -41,11 +43,11 @@ public class ExchangeRate {
     this.targetCurrencyId = targetCurrencyId;
   }
 
-  public double getRate() {
+  public BigDecimal getRate() {
     return rate;
   }
 
-  public void setRate(double rate) {
+  public void setRate(BigDecimal rate) {
     this.rate = rate;
   }
 }
